requirements.txt
streamlit>=1.28.0
requests>=2.28.0
psycopg2-binary>=2.9.0
